+++ src/screens/hardware-wallets/PairHardwareWalletAgainSheet.tsx
@@ -17,7 +17,7 @@ import Animated, {
   withSequence,
   withTiming,
 } from 'react-native-reanimated';
-import { CancelButton } from '@/screens/hardware-wallets/components/CancelButton';
+// import { CancelButton } from '@/screens/hardware-wallets/components/CancelButton';
 import gridDotsLight from '@/assets/dot-grid-light.png';
 import gridDotsDark from '@/assets/dot-grid-dark.png';
 import { useTheme } from '@/theme';
@@ -136,7 +136,6 @@ export const PairHardwareWalletAgainSheet = () => {
           </ImgixImage>
         </ImgixImage>
       </Box>
-      <CancelButton />
     </Layout>
   );
 };
